Extract generating ids out of OtpTransitServiceBuilder and test.
